+++ scripts/tools/obstacle.js
@@ -1,6 +1,7 @@
 // it's nothing, planned to expand further for obstacle related
 // TODO: crouch wall detection
 // TODO: <15ms wall detection
+// TODO: 3-wide wall detection
 function countInteractiveObstacle(obstacles) {
     let count = 0;
     for (let i = obstacles.length - 1; i >= 0; i--)